fixing java14 test compile.
git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@2090 696c4f5c-992c-4387-df9e-f6c076136008

+++ test/ca/odell/glazedlists/matchers/ThresholdMatcherEditorTest.java
@@ -440,7 +440,7 @@ public class ThresholdMatcherEditorTest extends TestCase {
      */
     private static class AbsComparator implements Comparator<Integer> {
         public int compare(Integer o1, Integer o2) {
-            return Math.abs(o1) - Math.abs(o2);
+            return Math.abs(o1.intValue()) - Math.abs(o2.intValue());
         }
     }
 }
